# app/Http/Resources/WorkerHashRateResource.php
<?php

declare(strict_types=1);

namespace App\Http\Resources;

use App\Models\WorkerHashrate;
use Illuminate\Http\Resources\Json\JsonResource;

/**
 * @see WorkerHashrate
 *
 * @OA\Schema(
 * schema="WorkerHashRateResource",
 * type="object",
 * @OA\Property(property="id", type="integer"),
 * @OA\Property(property="hash", type="number", format="float"),
 * @OA\Property(property="unit", type="string"),
 * @OA\Property(property="worker_id", type="integer"),
 * )
 */
class WorkerHashRateResource extends JsonResource
{
    public function toArray($request): array
    {
        return [
            'id' => $this->id,
            'hash' => $this->hash,
            'unit' => $this->unit,
            'worker_id' => $this->worker_id,
        ];
    }
}