# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allbtc Swagger UI",
        "description": "Документация OpenApi для Allbtc pool",
        "contact": {
            "email": "ognevoy.bima@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/v1/"
        }
    ],
    "paths": {
...
        "/workerhashrate/{worker}": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Get hash rates for a worker",
                "parameters": [
                    {
                        "name": "worker",
                        "in": "path",
                        "description": "Worker's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of hash rates to retrieve (default: 24)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkerHashRateResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Worker not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "WorkerHashRateResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "hash": {
                        "type": "number",
                        "format": "float"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "worker_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}